+++ packaging/homebrew/tardigrade.rb
@@ -41,11 +41,12 @@ class Tardigrade < Formula
   end
 
   def install
-    bin.install "tardigrade"
+    bin.install "tardi"
+    bin.install_symlink bin/"tardi" => "tardigrade"
   end
 
   service do
-    run [opt_bin/"tardigrade"]
+    run [opt_bin/"tardi"]
     keep_alive true
     log_path var/"log/tardigrade.log"
     error_log_path var/"log/tardigrade.log"
@@ -53,10 +54,7 @@ class Tardigrade < Formula
   end
 
   test do
-    pid = fork do
-      exec bin/"tardigrade", "--version"
-    end
-    Process.wait pid
-    assert_equal 0, $CHILD_STATUS.exitstatus
+    system bin/"tardi", "version"
+    system bin/"tardigrade", "version"
   end
 end